some mumctl -> mum::cli
